Add cli unit test
